fix: review refactor
