+++ feat-hazard-perception/src/components/hazard-perception-video-detail/hazard-perception-video-detail.scss
