+++ feat-hazard-perception/src/components/hazard-perception-video-result/hazard-perception-video-result.scss
@@ -1,4 +1,4 @@
-.hazard-perception-video-finished {
+.hazard-perception-video-result {
     width: 100%;
     height: calc(100vh - 60px);
     display: flex;
