+++ ui/src/index.ts
@@ -15,6 +15,3 @@ export * from './tabs/tabs';
 export * from './test-progress/test-progress-card';
 export * from './test-progress/test-progress-list';
 export * from './text-route-link/text-route-link';
-export * from './hazard-perception-video-card/hazard-perception-video-card';
-export * from './hazard-perception-video-detail/hazard-perception-video-detail';
-export * from './hazard-perception-video-finished/hazard-perception-video-finished';
